% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_filter.R
\name{feature_filter}
\alias{feature_filter}
\title{Filter out features with low counts}
\usage{
feature_filter(rawcounts, n_samples)
}
\arguments{
\item{rawcounts}{counts matrix}

\item{n_samples}{minimum number of samples passing the filter}
}
\value{
logical vector
}
\description{
Uses sequence read counts and number of samples per group to filter
out genes with low number of read counts in most samples.
}
